+++ betty/console/project.py
@@ -19,7 +19,7 @@ from betty.locale.localizable.gettext import _
 from betty.locale.localizable.markup import AnyEnumeration
 from betty.portable.file import assert_load_file
 from betty.project import Project, ProjectData
-from betty.serde import Serializer
+from betty.serialize import Serializer
 from betty.user import User
