+++ betty/tests/commands/test_config.py
@@ -4,8 +4,8 @@ from babel import Locale
 from pytest_mock import MockerFixture
 
 from betty.app import App, AppData
-from betty.plugins.serializer.json import Json
 from betty.portable.file import assert_load_file
+from betty.serializers.json import Json
 from betty.test_utils.console import run
