+++ betty/tests/commands/test_new.py
@@ -6,9 +6,9 @@ from betty.data import Data
 from betty.loaders.gramps import Gramps, GrampsData
 from betty.locale import DEFAULT_LOCALE_TAG, to_language_tag
 from betty.locale.localize import DEFAULT_LOCALIZER
-from betty.plugins.serializer.json import Json
 from betty.portable.file import assert_load_file
 from betty.project import ProjectData
+from betty.serializers.json import Json
 from betty.test_utils.conftest import IsolatedAppFactory
 from betty.test_utils.console import run
 from betty.test_utils.user import StaticUser
