+++ betty/tests/project/test_new.py
@@ -2,10 +2,10 @@ from pathlib import Path
 
 from betty.app import App
 from betty.locale.localize import DEFAULT_LOCALIZER
-from betty.plugins.serializer.json import Json
 from betty.portable.file import assert_load_file
 from betty.project import ProjectData
 from betty.project.new import new
+from betty.serializers.json import Json
 
 
 async def test_new(isolated_app: App, tmp_path: Path) -> None:
