+++ pyproject.toml
@@ -100,7 +100,7 @@ betty = 'betty.console:main_standalone'
 'place-type' = 'betty.place_type:PlaceTypeDefinition'
 'renderer' = 'betty.render:RendererDefinition'
 'role' = 'betty.role:RoleDefinition'
-'serializer' = 'betty.serde:SerializerDefinition'
+'serializer' = 'betty.serialize:SerializerDefinition'
 'server' = 'betty.server:ServerDefinition'
 
 [project.entry-points.'betty.asset_directory']
@@ -301,8 +301,8 @@ betty = 'betty.console:main_standalone'
 'wikipedia-rest-api' = 'betty.plugins.rate_limit.wikipedia_rest_api:WIKIPEDIA_REST_API'
 
 [project.entry-points.'betty.serializer']
-'json' = 'betty.plugins.serializer.json:Json'
-'yaml' = 'betty.plugins.serializer.yaml:Yaml'
+'json' = 'betty.serializers.json:Json'
+'yaml' = 'betty.serializers.yaml:Yaml'
 
 [project.entry-points.'betty.server']
 'builtin' = 'betty.servers.project_builtin:ProjectBuiltinServer'
